Add metadce test for minimal threaded program
This sets up a baseline for the size of mutli-threads
program which we can then work to improve on and also
prevents regressions.
