+++ tests/other/metadce/minimal_Oz_USE_PTHREADS_PROXY_TO_PTHREAD.exports
@@ -1,0 +1,33 @@
+A
+B
+C
+D
+E
+F
+G
+H
+I
+J
+K
+L
+M
+N
+O
+P
+Q
+R
+S
+T
+U
+V
+W
+X
+Y
+s
+t
+u
+v
+w
+x
+y
+z
