+++ tests/other/metadce/minimal_Oz_USE_PTHREADS_PROXY_TO_PTHREAD.funcs
@@ -1,0 +1,82 @@
+$GetOrAllocateQueue
+$GetQueue
+$__ashlti3
+$__emscripten_pthread_data_constructor
+$__errno_location
+$__fwritex
+$__lshrti3
+$__pthread_mutex_lock
+$__pthread_mutex_timedlock
+$__pthread_mutex_trylock
+$__pthread_mutex_trylock_owner
+$__pthread_mutex_unlock
+$__pthread_self_internal
+$__pthread_setcancelstate
+$__pthread_tsd_run_dtors
+$__timedwait
+$__timedwait_cp
+$__towrite
+$__trunctfdf2
+$__vfprintf_internal
+$__wasm_call_ctors
+$__wasm_init_memory
+$_do_call
+$_emscripten_call_on_thread
+$_emscripten_thread_init
+$_main_thread
+$a_cas
+$add
+$dispose_chunk
+$dlfree
+$dlmalloc
+$dlmemalign
+$do_emscripten_dispatch_to_thread
+$em_queued_call_free
+$em_queued_call_malloc
+$emscripten_async_run_in_main_thread
+$emscripten_atomic_store_u32
+$emscripten_current_thread_process_queued_calls
+$emscripten_get_global_libc
+$emscripten_main_browser_thread_id
+$emscripten_main_thread_process_queued_calls
+$emscripten_register_main_browser_thread_id
+$emscripten_run_in_main_runtime_thread_js
+$emscripten_stack_set_limits
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread
+$emscripten_sync_run_in_main_thread_xprintf_varargs
+$emscripten_tls_init
+$emscripten_wait_for_call_v
+$fmt_fp
+$fmt_o
+$fmt_u
+$fmt_x
+$frexp
+$getint
+$init_mparams
+$internal_memalign
+$isdigit
+$memchr
+$memcpy
+$memset
+$out
+$pad
+$pop_arg
+$pop_arg_long_double
+$printf_core
+$pthread_mutex_init
+$sbrk
+$sn_write
+$stackAlloc
+$stackRestore
+$stackSave
+$vsnprintf
+$wcrtomb
+$wctomb
