+++ tests/other/metadce/minimal_Oz_USE_PTHREADS_PROXY_TO_PTHREAD.imports
@@ -1,0 +1,18 @@
+a.a
+a.b
+a.c
+a.d
+a.e
+a.f
+a.g
+a.h
+a.i
+a.j
+a.k
+a.l
+a.m
+a.n
+a.o
+a.p
+a.q
+a.r
