+++ tests/other/metadce/minimal_Oz_USE_PTHREADS_PROXY_TO_PTHREAD.sent
@@ -1,0 +1,18 @@
+a
+b
+c
+d
+e
+f
+g
+h
+i
+j
+k
+l
+m
+n
+o
+p
+q
+r
